renamed StoreRetrieveService to TrainerService
